Upgrade PDFBox API to 2.0.26

+++ library/src/main/java/com/tom_roush/pdfbox/pdmodel/graphics/image/CCITTFactory.java
@@ -306,7 +306,7 @@ public final class CCITTFactory
             }
 
             // Relocate to the first set of tags
-            int address = readlong(endianess, reader);
+            long address = readlong(endianess, reader);
             reader.seek(address);
 
             // If some higher page number is required, skip this page's tags, 
@@ -318,7 +318,7 @@ public final class CCITTFactory
                 {
                     throw new IOException("Not a valid tiff file");
                 }
-                reader.seek(address + 2 + numtags * 12);
+                reader.seek(address + 2 + numtags * 12L);
                 address = readlong(endianess, reader);
                 if (address == 0)
                 {
